Major Change to Choice / Task Design
I am now changing the task design (again). Here is my email explanation to Stefan on July 14: 

I'm thinking I'll still have two options, that draw from two different sets of payouts. So for example option A might cost 90 Credits and randomly result in one of {0, 30, 60 , 150, 180, 500} where option B might also cost 90 credits but draw from {0, 40, 50 , 160, 170, 500}... or something like that with both options balanced for long-run EV. This will double the number of LDWs that occur and substantially increase our power for the EEG analysis. We lose behavioural preference as a dependent variable, but would still be able to fit a learning model to the data if we wanted to look at behaviour.

To be clear I have in mind a design that has no difference between the two choice conditions except the 5 payout amounts, and even these will be relatively similar. The rate of LDWs/true wins/losses will be the same across both conditions. The 2 choice structure will just be a cover story to keep them thinking about outcomes while decision making, so that the EEG data we have is related to outcome processing and decision-making.